# api/openapi.yaml
openapi: "3.0.3"
info:
  title: OneStepTwo API
  version: "0.1.0"
paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: Health check
      responses:
        "200":
          description: Service is healthy
  /v1/children:
    get:
      operationId: getV1Children
      summary: List the caller's active-org children
      responses:
        "200":
          description: Children for the caller's active organization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChildResponse'
        "401":
          description: Unauthorized
        "403":
          description: No active organization in session
    post:
      operationId: postV1Children
      summary: Create a child profile with parental consent (atomic consent gate)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        "201":
          description: Child profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildResponse'
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        "401":
          description: Unauthorized
  /v1/children/{id}:
    delete:
      operationId: deleteV1ChildrenId
      summary: Delete a child profile and all associated data (GDPR/COPPA erasure)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Child profile deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErasureConfirmation'
        "401":
          description: Unauthorized
        "404":
          description: Child not found
  /v1/account:
    delete:
      operationId: deleteV1Account
      summary: Delete the entire account and all associated data (GDPR/COPPA erasure)
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErasureConfirmation'
        "401":
          description: Unauthorized
components:
  schemas:
    CreateChildRequest:
      type: object
      required:
        - nickname
        - birth_month
        - birth_year
        - consent
      properties:
        nickname:
          type: string
        birth_month:
          type: integer
        birth_year:
          type: integer
        consent:
          type: object
          required:
            - app_version
            - consent_text_version
          properties:
            app_version:
              type: string
            consent_text_version:
              type: string
    ChildResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clerk_org_id:
          type: string
        nickname:
          type: string
        birth_month:
          type: integer
        birth_year:
          type: integer
    ErasureConfirmation:
      type: object
      properties:
        deleted_children:
          type: integer
        deleted_events:
          type: integer
        deleted_consent_events:
          type: integer
        deleted_notification_preferences:
          type: integer
        deleted_device_tokens:
          type: integer
        requested_by:
          type: string
        requested_at:
          type: string
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string